Separate "test files" from "files to load for each test"
The `--file` argument is used for configuring files that need to be loaded as part of running any
set of tests, and these are intended to add global hooks etc.

For that reason these need to be kept separate (in the mochaOpts) from the actual test files, and the
worker needs to explicitly load them when running a set of tests.